requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-jose[cryptography]
passlib
bcrypt==4.0.1
pandas
aiofiles
pydantic-settings
email-validator
